requirements.txt
aiogram==3.24.0
openpyxl==3.1.5 
python-dotenv==1.2.1